+++ src/checkers/package/custom.rs
@@ -42,3 +42,7 @@ pub fn is_installed(package: &CustomInstaller) -> Result<bool, CheckError> {
 
     Ok(stdout.contains(package.version.as_str()))
 }
+
+pub fn is_upgradable(package: &CustomInstaller) -> Result<bool, CheckError> {
+    Ok(true)
+}